+++ webknossos/tests/test_skeleton.py
@@ -265,7 +265,7 @@ def test_simple_initialization_and_representations(tmp_upath: UPath) -> None:
         )
     assert nml == wk.Skeleton.load(nml_path)
     assert str(nml) == (
-        "Skeleton(_child_groups=<No child groups>, _child_trees=<No child trees>, _voxel_size=VoxelSize(factor=(0.5, 0.5, 0.5), unit=<LengthUnit.NANOMETER: 'nanometer'>), dataset_name='ds_name', dataset_id=None, organization_id=None, description=None)"
+        "Skeleton(_child_groups=<No child groups>, _child_trees=<No child trees>, _voxel_size=VoxelSize(factor=Vec3Float(0.5,0.5,0.5), unit=<LengthUnit.NANOMETER: 'nanometer'>), dataset_name='ds_name', dataset_id=None, organization_id=None, description=None)"
     )
 
     my_group = nml.add_group("my_group")
@@ -315,7 +315,7 @@ def test_simple_initialization_and_representations(tmp_upath: UPath) -> None:
         )
     assert nml == wk.Skeleton.load(nml_path)
     assert str(nml) == (
-        "Skeleton(_child_groups=<1 child group>, _child_trees=<1 child tree>, _voxel_size=VoxelSize(factor=(0.5, 0.5, 0.5), unit=<LengthUnit.NANOMETER: 'nanometer'>), dataset_name='ds_name', dataset_id=None, organization_id=None, description=None)"
+        "Skeleton(_child_groups=<1 child group>, _child_trees=<1 child tree>, _voxel_size=VoxelSize(factor=Vec3Float(0.5,0.5,0.5), unit=<LengthUnit.NANOMETER: 'nanometer'>), dataset_name='ds_name', dataset_id=None, organization_id=None, description=None)"
     )
     assert (
         str(my_group)